review and update existing content
- Move snippets
- edit for clarity
- markdown lint
