+++ docs/csharp/programming-guide/exceptions/how-to-handle-an-exception-using-try-catch.md
@@ -1,7 +1,7 @@
 ---
 title: "How to handle an exception using try-catch - C# Programming Guide"
 description: Learn how to handle an exception using a try-catch block. See a code example and view additional available resources.
-ms.date: 07/20/2015
+ms.date: 12/09/2020
 helpviewer_keywords: 
   - "exception handling [C#], try/catch blocks"
   - "exceptions [C#], try/catch blocks"
@@ -10,20 +10,14 @@ ms.assetid: ca8e3773-980e-4767-8633-7408540e9818
 ---
 # How to handle an exception using try/catch (C# Programming Guide)
 
-The purpose of a [try-catch](../../language-reference/keywords/try-catch.md) block is to catch and handle an exception generated by working code. Some exceptions can be handled in a `catch` block and the problem solved without the exception being re-thrown; however, more often the only thing that you can do is make sure that the appropriate exception is thrown.  
-  
-## Example  
+The purpose of a [try-catch](../../language-reference/keywords/try-catch.md) block is to catch and handle an exception generated by working code. Some exceptions can be handled in a `catch` block and the problem solved without the exception being rethrown; however, more often the only thing that you can do is make sure that the appropriate exception is thrown.
 
- In this example, <xref:System.IndexOutOfRangeException> is not the most appropriate exception: <xref:System.ArgumentOutOfRangeException> makes more sense for the method because the error is caused by the `index` argument passed in by the caller.  
-  
- [!code-csharp[csProgGuideExceptions#5](~/samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideExceptions/CS/Exceptions.cs#5)]  
-  
-## Comments  
+## Example
 
- The code that causes an exception is enclosed in the `try` block. A `catch` statement is added immediately after to handle `IndexOutOfRangeException`, if it occurs. The `catch` block handles the `IndexOutOfRangeException` and throws the more appropriate `ArgumentOutOfRangeException` exception instead. In order to provide the caller with as much information as possible, consider specifying the original exception as the <xref:System.Exception.InnerException%2A> of the new exception. Because the <xref:System.Exception.InnerException%2A> property is [read-only](../../properties.md#read-only), you must assign it in the constructor of the new exception.  
-  
-## See also
+In this example, <xref:System.IndexOutOfRangeException> isn't the most appropriate exception: <xref:System.ArgumentOutOfRangeException> makes more sense for the method because the error is caused by the `index` argument passed in by the caller.
 
-- [C# Programming Guide](../index.md)
-- [Exceptions and Exception Handling](./index.md)
-- [Exception Handling](./exception-handling.md)
+:::code language="csharp" source="snippets/exceptions/ExampleTryCatch.cs" id="ExampleTryCatch":::
+
+## Comments
+
+The code that causes an exception is enclosed in the `try` block. A `catch` statement is added immediately after to handle `IndexOutOfRangeException`, if it occurs. The `catch` block handles the `IndexOutOfRangeException` and throws the more appropriate `ArgumentOutOfRangeException` exception instead. In order to provide the caller with as much information as possible, consider specifying the original exception as the <xref:System.Exception.InnerException%2A> of the new exception. Because the <xref:System.Exception.InnerException%2A> property is [read-only](../../properties.md#read-only), you must assign it in the constructor of the new exception.
