+++ docs/csharp/programming-guide/exceptions/index.md
@@ -1,7 +1,7 @@
 ---
 title: "Exceptions and Exception Handling - C# Programming Guide"
 description: Learn about exceptions and exception handling. These C# features help deal with unexpected or exceptional situations that happen when a program is running.
-ms.date: 07/20/2015
+ms.date: 12/09/2020
 helpviewer_keywords:
   - "exception handling [C#]"
   - "exceptions [C#]"
@@ -10,13 +10,13 @@ ms.assetid: 0001887f-4fa2-47e2-8034-2819477e2344
 ---
 # Exceptions and Exception Handling (C# Programming Guide)
 
-The C# language's exception handling features help you deal with any unexpected or exceptional situations that occur when a program is running. Exception handling uses the `try`, `catch`, and `finally` keywords to try actions that may not succeed, to handle failures when you decide that it is reasonable to do so, and to clean up resources afterward. Exceptions can be generated by the common language runtime (CLR), by .NET or third-party libraries, or by application code. Exceptions are created by using the `throw` keyword.
+The C# language's exception handling features help you deal with any unexpected or exceptional situations that occur when a program is running. Exception handling uses the `try`, `catch`, and `finally` keywords to try actions that may not succeed, to handle failures when you decide that it's reasonable to do so, and to clean up resources afterward. Exceptions can be generated by the common language runtime (CLR), by .NET or third-party libraries, or by application code. Exceptions are created by using the `throw` keyword.
 
-In many cases, an exception may be thrown not by a method that your code has called directly, but by another method further down in the call stack. When this happens, the CLR will unwind the stack, looking for a method with a `catch` block for the specific exception type, and it will execute the first such `catch` block that if finds. If it finds no appropriate `catch` block anywhere in the call stack, it will terminate the process and display a message to the user.
+In many cases, an exception may be thrown not by a method that your code has called directly, but by another method further down in the call stack. When an exception is thrown happens, the CLR will unwind the stack, looking for a method with a `catch` block for the specific exception type, and it will execute the first such `catch` block that if finds. If it finds no appropriate `catch` block anywhere in the call stack, it will terminate the process and display a message to the user.
 
 In this example, a method tests for division by zero and catches the error. Without the exception handling, this program would terminate with a **DivideByZeroException was unhandled** error.
 
-[!code-csharp[csProgGuideExceptions#18](~/samples/snippets/csharp/VS_Snippets_VBCSharp/csProgGuideExceptions/CS/Exceptions.cs#18)]
+:::code language="csharp" source="snippets/exceptions/ExceptionTest.cs" ID="ExceptionTest":::
 
 ## Exceptions Overview
 
@@ -26,24 +26,12 @@ Exceptions have the following properties:
 - Use a `try` block around the statements that might throw exceptions.
 - Once an exception occurs in the `try` block, the flow of control jumps to the first associated exception handler that is present anywhere in the call stack. In C#, the `catch` keyword is used to define an exception handler.
 - If no exception handler for a given exception is present, the program stops executing with an error message.
-- Do not catch an exception unless you can handle it and leave the application in a known state. If you catch `System.Exception`, re-throw it using the `throw` keyword at the end of the `catch` block.
+- Don't catch an exception unless you can handle it and leave the application in a known state. If you catch `System.Exception`, rethrow it using the `throw` keyword at the end of the `catch` block.
 - If a `catch` block defines an exception variable, you can use it to obtain more information about the type of exception that occurred.
 - Exceptions can be explicitly generated by a program by using the `throw` keyword.
 - Exception objects contain detailed information about the error, such as the state of the call stack and a text description of the error.
 - Code in a `finally` block is executed even if an exception is thrown. Use a `finally` block to release resources, for example to close any streams or files that were opened in the `try` block.
 - Managed exceptions in .NET are implemented on top of the Win32 structured exception handling mechanism. For more information, see [Structured Exception Handling (C/C++)](/cpp/cpp/structured-exception-handling-c-cpp) and [A Crash Course on the Depths of Win32 Structured Exception Handling](http://bytepointer.com/resources/pietrek_crash_course_depths_of_win32_seh.htm).
-
-## Related Sections
-
-See the following articles for more information about exceptions and exception handling:
-
-- [Using Exceptions](using-exceptions.md)
-- [Exception Handling](exception-handling.md)
-- [Creating and Throwing Exceptions](creating-and-throwing-exceptions.md)
-- [Compiler-Generated Exceptions](compiler-generated-exceptions.md)
-- [How to handle an exception using try/catch (C# Programming Guide)](how-to-handle-an-exception-using-try-catch.md)
-- [How to execute cleanup code using finally](how-to-execute-cleanup-code-using-finally.md)
-- [How to catch a non-CLS exception](how-to-catch-a-non-cls-exception.md)
 
 ## C# Language Specification
 
@@ -52,7 +40,6 @@ For more information, see [Exceptions](~/_csharplang/spec/exceptions.md) in the 
 ## See also
 
 - <xref:System.SystemException>
-- [C# Programming Guide](../index.md)
 - [C# Keywords](../../language-reference/keywords/index.md)
 - [throw](../../language-reference/keywords/throw.md)
 - [try-catch](../../language-reference/keywords/try-catch.md)
