+++ docs/csharp/programming-guide/exceptions/snippets/exceptions/exceptions.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+    <StartupObject>exceptions.Program</StartupObject>
+  </PropertyGroup>
+
+</Project>
